Renamed WgpuOffscreenCanvas to WgpuOffscreenCanvasBase
