+++ wgpu/gui/__init__.py
@@ -3,11 +3,11 @@ Code to provide a canvas to render to.
 """
 
 from .base import WgpuCanvasInterface, WgpuCanvasBase, WgpuAutoGui  # noqa: F401
-from ._offscreen import WgpuOffscreenCanvas  # noqa: F401
+from .offscreen import WgpuOffscreenCanvasBase  # noqa: F401
 
 __all__ = [
     "WgpuCanvasInterface",
     "WgpuCanvasBase",
     "WgpuAutoGui",
-    "WgpuOffscreenCanvas",
+    "WgpuOffscreenCanvasBase",
 ]
